+++ mayan/apps/common/http.py
@@ -6,14 +6,19 @@ from django.utils.six import PY3
 
 
 class URL(object):
-    def __init__(self, path=None, query_string=None):
+    def __init__(self, path=None, query_string=None, query=None):
         self._path = path
         self._query_string = query_string
+        self._query = query
+
         kwargs = {'mutable': True}
-        if query_string:
-            kwargs['query_string'] = query_string.encode('utf-8')
+        if self._query_string:
+            kwargs['query_string'] = self._query_string.encode('utf-8')
 
         self._args = QueryDict(**kwargs)
+
+        if self._query:
+            self.args.update(self._query)
 
     @property
     def args(self):